Model downloading (#69)
Change the process for discovering releases from polling the github API for release information to using a JSON index file stored in the python package. The index contains model names, versions, license information and direct download URLs. This means we no longer need to authenticate API requests to check for models.

Now checks if models are downloaded before downloading. Also add a force argument to re-download if desired.

The idea would be to put out new models with a new zip structure (a bare folder rather than nested). The current version on PyPi would continue to work, and future versions can have models added to the index JSON and should work. Users wanting new versions of models for an older instanseg version can download them and install away.

Usage will be as before, but now with an extra version argument. name and version are used in string matching. Models are stored in the same place as before, but now with a subdirectory within each model for every version. We assume the same is true for user models, although we do print a message informing them of that.